Create a small set of predefined checks & helpers.
This brings another library target, libatcheck.a and it's accompanying
atcheck.h file.

The checks are only a few so far: a bunch of eq_<type>
checks that so far are only capable of handling integer types and
pointers.

Keep in mind that, along with the checks, this library brings a few
functions (at_allocf, at_asprintf, at_vasprintf) that may or may not be
enabled depending on a few build time flags:

at_allocf and at_asprintf are always enabled, but they may delegate the
heavy work to the implementation's vasprintf if the HAS_VASPRINTF flag
is enabled.

at_vasprintf depends on the implementation having va_copy available. You
may signal its availability by setting the HAS_VA_COPY flag on build.
Keep in mind that it will only build correctly (in the presence of
HAS_VA_COPY) if the va_copy macro is declared on stdarg.h during regular
build. If that's not the case, other work may be needed to enable this.